Refactoring to accommodate for notebook
Set up code in preparation to create notebook
