+++ MainGUI.py
@@ -4,27 +4,20 @@ from tkinter import PhotoImage
 
 import MainPageObjects
 import TimerObject
+import TimerFrameObject
 
-backgroundColor = "#D1FFB0"
+backgroundColor = "#3A7069"
 
 TK_SILENCE_DEPRECATION=1 #suppress warnings
 
-class MainFrame:
+class MainFrame(tk.Tk):
     def __init__(self):
-        self.root = tk.Tk() #creates main root
-        self.root.config(bg= backgroundColor)
-        self.root.title("Track Timer 1.0.0") #title display at the top
-        self.root.geometry('500x700') #inital dimensions of window
+        #self.root = ?
+        container = tk.Frame(self)
 
-        #image of clock    
-        self.image1 = Canvas(self.root, width =350, height=350)
-        self.image1.place(relx=.5, rely=.4, anchor="center")
-        self.photo1 = PhotoImage(file='Images/Clock2-resized.gif') #TODO: Figure out how to make corners transparent
-        self.image1.create_image((1, 1), anchor='nw',image= self.photo1)
-
-main = MainFrame()
-mainLabel = MainPageObjects.StartingLabel(main.root)
-timer = TimerObject.Timer(main.root)
-goalDashboardButton = MainPageObjects.GoalDashboardButton(main.root)
+noteBock = MainFrame()
+TimerFrame = TimerFrameObject.TimerFrame(MainFrame.root)
+timer = TimerObject.Timer(MainFrame.root)
+goalDashboardButton = MainPageObjects.GoalDashboardButton(MainFrame.root)
 
 main.root.mainloop()
