+++ MainPageObjects.py
@@ -1,13 +1,6 @@
 import tkinter as tk
 
-backgroundColor = "#D1FFB0" #background color global for main theme
-
-class StartingLabel:
-    def __init__(self, root):
-        self.root = root
-        self.label = tk.Label(self.root, text = "Click start to begin", font=('MS Sans Serif', 20), bg= backgroundColor)
-        self.label.place(relx=.5, rely=.1, anchor="center")
-
+backgroundColor = "#3A7069" #background color global for main theme
 
 class TimeTags:
     def __init__(self, root):
@@ -17,7 +10,7 @@ class TimeTags:
             usr = tk.Entry(self.root, relief='flat', bg = backgroundColor)   # store user input into a database
             canvas.create_window(200, 250, window=usr)
             
-        tagBtn = tk.Button(self.root, text = "Create a Tag", bg = backgroundColor, state = 'active', command = userin)
+        tagBtn = tk.Button(self.root, text = "Create a Tag", bg =backgroundColor, state = 'active', command = userin)
         tagBtn.place(x = 250, y = 100)
 
 class GoalDashboardButton:
