+++ TimerObject.py
@@ -2,8 +2,7 @@ import tkinter as tk
 import time
 import threading
 
-backgroundColor = "#D1FFB0"
-
+backgroundColor = "#3A7069"
 
 #Tutorial referenced: https://www.youtube.com/watch?v=iP7CaRg9OPA
 class Timer:
@@ -26,21 +25,13 @@ class Timer:
         currentHour = currentMinute // 60
         return f"{int(currentHour):02d}:{int(currentMinute):02d}:{int(currentSecond):02d}"
 
+
     def startStop(self):
         self.currentlyRunning = not self.currentlyRunning
 
         if(self.currentlyRunning):
             self.button.config(text = "Stop") #change button label
             threading.Thread(target=self.countUp).start() #start new thread to count
-
-            #Sorry linnea i dont think we need this any more :(
-            # stopTimeAsString = 0
-            # stopTimeAsString = now.strftime("%H%M") # Attempting to capture stop time when "stop" is clicked - LPC
-            # stopTime = int(stopTimeAsString)
-
-            # if stopTime != 0:
-            #     timeElapsed = (stopTime - startTime) # Might actually need to convert these from strings to ints? - LPC
-            #     # timer.label = timeElapsed TODO: Change label to show time elapsed rather than 00:00
 
         else:
             self.button.config(text = "Start") #change button label
